+++ src/simmate/utilities/__init__.py
@@ -1,6 +1,12 @@
 # -*- coding: utf-8 -*-
 
-from .files import get_directory, make_archive, archive_old_runs, empty_directory
+from .files import (
+    get_directory,
+    make_archive,
+    make_error_archive,
+    archive_old_runs,
+    empty_directory,
+)
 from .other import get_conda_env, get_doc_from_readme, get_chemical_subsystems
 
 # must be last to prevent circular import